+++ cypress/e2e/demo.cy.js
@@ -1,5 +1,6 @@
 context("WM", () => {
   const screenshotConfig = { blackout: ["ol[role='tree']"] };
+  const defaultWait = 800;
 
   beforeEach(() => {
     cy.visit("iframe.html?args=&id=demo--defult&viewMode=story");
@@ -14,7 +15,6 @@ context("WM", () => {
   it("should maximize/restore window", () => {
     cy.contains("with title").click();
     cy.get("[data-testid='window-frame']").should("be.visible").as("window");
-    cy.get("@window").contains("scroll to bottom").should("be.enabled");
     cy.document().matchImage({ screenshotConfig });
     cy.get("@window").contains("with title").dblclick();
     cy.document().matchImage({ screenshotConfig });
@@ -26,14 +26,14 @@ context("WM", () => {
   it("should prevent window from going offscreen", () => {
     cy.contains("danger").click();
     cy.get("[data-testid='window-frame']").should("be.visible").as("window");
-    cy.contains("to bottom").should("be.enabled");
+    cy.contains("to bottom").should("be.disabled");
     cy.document().matchImage({ screenshotConfig });
     cy.viewport(500, 500);
-    cy.wait(300);
-    cy.document().matchImage({ screenshotConfig });
+    cy.wait(defaultWait);
+    cy.get("[data-testid='window-frame']").matchImage({ screenshotConfig });
     cy.contains("to bottom").should("be.enabled");
-    cy.viewport("macbook-13");
-    cy.wait(300);
+    cy.viewport(1280, 720);
+    cy.wait(defaultWait);
     cy.contains("to bottom").should("be.disabled");
     cy.contains("add line").should("be.visible").click().click().click();
     cy.contains("to bottom").should("be.enabled");
@@ -43,14 +43,14 @@ context("WM", () => {
   it("should prevent window from going offscreen (minimal size)", () => {
     cy.contains("minimal size").click();
     cy.get("[data-testid='window-frame']").should("be.visible").as("window");
-    cy.contains("to bottom").should("be.enabled");
+    cy.contains("to bottom").should("be.disabled");
     cy.document().matchImage({ screenshotConfig });
     cy.viewport(500, 500);
-    cy.wait(300);
-    cy.document().matchImage({ screenshotConfig });
+    cy.wait(defaultWait);
+    cy.get("[data-testid='window-frame']").matchImage({ screenshotConfig });
     cy.contains("to bottom").should("be.enabled");
-    cy.viewport("macbook-13");
-    cy.wait(300);
+    cy.viewport(1280, 720);
+    cy.wait(defaultWait);
     cy.contains("to bottom").should("be.disabled");
     cy.contains("add line").should("be.visible").click().click().click();
     cy.contains("to bottom").should("be.enabled");
@@ -65,6 +65,20 @@ context("WM", () => {
         const { top, left } = $el[0].getBoundingClientRect();
         expect(top).to.eq(Math.round(top));
         expect(left).to.eq(Math.round(left));
+      });
+  });
+
+  it("should respect layoutData position", () => {
+    cy.viewport(1000, 1000);
+    cy.contains("left").click();
+    cy.get("[data-testid='window-frame']")
+      .should("be.visible")
+      .should(($el) => {
+        const { top, left, right, bottom } = $el[0].getBoundingClientRect();
+        expect(top).to.eq(50);
+        expect(left).to.eq(200);
+        expect(right).to.eq(950);
+        expect(bottom).to.eq(800);
       });
   });
 });