+++ src/components/WindowsViewport.tsx
@@ -2,7 +2,6 @@ import { css, cx } from "@emotion/css";
 import { useTheme } from "@emotion/react";
 import React, { PropsWithChildren, useMemo } from "react";
 import { useSize } from "../hooks";
-import { useVisualHeight } from "../hooks/useVisualHeight";
 import { ViewportContext } from "../ViewportContext";
 
 const ignorePointerEvents = css({
@@ -21,14 +20,12 @@ const fullscreenFixed = css({
 });
 
 export function WindowsViewport({ children }: PropsWithChildren<unknown>): JSX.Element {
-  const { observe, height, width, entry } = useSize();
-  const visualHeight = useVisualHeight();
-
-  const value = useMemo(() => ({ width, height: Math.min(height, visualHeight) }), [height, width, visualHeight]);
+  const { observe, height, width } = useSize();
+  const value = useMemo(() => ({ width, height }), [height, width]);
 
   const { zIndex } = useTheme();
   return (
-    <div className={cx(fullscreenFixed, ignorePointerEvents, css({ zIndex }))} ref={observe}>
+    <div id={"windowsViewport"} className={cx(fullscreenFixed, ignorePointerEvents, css({ zIndex }))} ref={observe}>
       <ViewportContext.Provider value={value}>{children}</ViewportContext.Provider>
     </div>
   );