+++ src/components/window/header/HeaderButtonZoom.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useRef } from "react";
+import React, { UIEvent, useCallback, useRef } from "react";
 import { HeaderButton } from "./HeaderButton";
 import RestoreIcon from "./restore.svg";
 import ZoomIcon from "./zoom.svg";
@@ -11,15 +11,20 @@ export interface HeaderButtonZoomProps {
 
 export function HeaderButtonZoom({ zoomDialog, isMaximized, keepFocus }: HeaderButtonZoomProps): JSX.Element {
   const ref = useRef<HTMLButtonElement>();
-  const onPointerDown = useCallback(() => {
-    zoomDialog();
-    if (!keepFocus) {
-      ref.current?.blur();
-    }
-  }, [keepFocus, zoomDialog]);
+  const action = useCallback(
+    (event: UIEvent) => {
+      event.preventDefault();
+
+      zoomDialog();
+      if (!keepFocus) {
+        ref.current?.blur();
+      }
+    },
+    [keepFocus, zoomDialog],
+  );
 
   return (
-    <HeaderButton name="zoom" ref={ref} onPointerDown={onPointerDown}>
+    <HeaderButton name="zoom" ref={ref} action={action}>
       {isMaximized ? <RestoreIcon /> : <ZoomIcon />}
     </HeaderButton>
   );