+++ src/demo/DebugButtons.tsx
@@ -1,12 +1,12 @@
 import React, { useMemo } from "react";
 import { useWindowManager } from "../hooks";
-import { WindowId } from "../types";
+import { WindowId, WindowType } from "../types";
 import { DemoWindowKind } from "./DemoWindowKind";
 
 export function DebugButtons({ currentId }: { currentId?: WindowId }): JSX.Element {
   const { open, close, closeAll } = useWindowManager<DemoWindowKind | string>(currentId);
 
-  const buttons = useMemo(
+  const buttons: Partial<WindowType>[] = useMemo(
     () => [
       undefined,
       { title: "with title" },
@@ -16,8 +16,9 @@ export function DebugButtons({ currentId }: { currentId?: WindowId }): JSX.Eleme
       { title: "danger, non-resizable", kind: DemoWindowKind.danger, isResizable: false },
       { title: "not modal", isModal: false },
       { title: "lazy loaded content", kind: DemoWindowKind.lazy },
-      { title: "initial size", kind: DemoWindowKind.lazy, width: 700, height: 700 },
-      { title: "minimal size (overflow fixed)", kind: DemoWindowKind.danger, width: 7000, height: 7000 },
+      { title: "initial size", kind: DemoWindowKind.lazy, layoutData: { width: 700, height: 700 } },
+      { title: "minimal size (overflow fixed)", kind: DemoWindowKind.danger, layoutData: { width: 7000, height: 7000 } },
+      { title: `initial position`, layoutData: { top: 50, left: 200, right: 50, bottom: 200 } },
       { title: `fixed id`, id: "fixed", parent: null },
     ],
     [],
@@ -26,11 +27,11 @@ export function DebugButtons({ currentId }: { currentId?: WindowId }): JSX.Eleme
   return (
     <div>
       {buttons.map((props, index) => (
-        <div key={index}>
+        <span key={index}>
           <button onClick={() => open(props)} style={{ fontWeight: "bold", color: "black", margin: ".5em" }}>
             add({JSON.stringify(props)})
           </button>
-        </div>
+        </span>
       ))}
       <button style={{ margin: ".5em" }} onClick={() => closeAll()}>
         close all